some minor corrections to the tests
Update AsyncKafkaConsumerTest.java

reverting unintentional changes
