+++ core/src/test/scala/integration/kafka/api/PlaintextConsumerCommitTest.scala
@@ -25,7 +25,7 @@ import org.junit.jupiter.params.provider.{Arguments, MethodSource}
 
 import java.time.Duration
 import java.util
-import java.util.{Collections, Optional}
+import java.util.Optional
 import java.util.stream.Stream
 import scala.jdk.CollectionConverters._
 
@@ -302,60 +302,6 @@ class PlaintextConsumerCommitTest extends AbstractConsumerTest {
     val otherConsumer = createConsumer()
     otherConsumer.assign(List(tp).asJava)
     consumeAndVerifyRecords(consumer = otherConsumer, numRecords = 1, startingOffset = 5, startingTimestamp = startingTimestamp)
-  }
-
-  @ParameterizedTest(name = TestInfoUtils.TestWithParameterizedQuorumAndGroupProtocolNames)
-  @MethodSource(Array("getTestQuorumAndGroupProtocolParametersAll"))
-  def testEndOffsets(quorum: String, groupProtocol: String): Unit = {
-    val producer = createProducer()
-    val startingTimestamp = System.currentTimeMillis()
-    val numRecords = 10000
-    (0 until numRecords).map { i =>
-      val timestamp = startingTimestamp + i.toLong
-      val record = new ProducerRecord(tp.topic(), tp.partition(), timestamp, s"key $i".getBytes, s"value $i".getBytes)
-      producer.send(record)
-      record
-    }
-    producer.flush()
-
-    val consumer = createConsumer()
-    consumer.subscribe(List(topic).asJava)
-    awaitAssignment(consumer, Set(tp, tp2))
-
-    val endOffsets = consumer.endOffsets(Set(tp).asJava)
-    assertEquals(numRecords, endOffsets.get(tp))
-  }
-
-  @ParameterizedTest(name = TestInfoUtils.TestWithParameterizedQuorumAndGroupProtocolNames)
-  @MethodSource(Array("getTestQuorumAndGroupProtocolParametersAll"))
-  def testFetchOffsetsForTime(quorum: String, groupProtocol: String): Unit = {
-    val numPartitions = 2
-    val producer = createProducer()
-    val timestampsToSearch = new util.HashMap[TopicPartition, java.lang.Long]()
-    var i = 0
-    for (part <- 0 until numPartitions) {
-      val tp = new TopicPartition(topic, part)
-      // key, val, and timestamp equal to the sequence number.
-      sendRecords(producer, numRecords = 100, tp, startingTimestamp = 0)
-      timestampsToSearch.put(tp, (i * 20).toLong)
-      i += 1
-    }
-
-    val consumer = createConsumer()
-    // Test negative target time
-    assertThrows(classOf[IllegalArgumentException],
-      () => consumer.offsetsForTimes(Collections.singletonMap(new TopicPartition(topic, 0), -1)))
-    val timestampOffsets = consumer.offsetsForTimes(timestampsToSearch)
-
-    val timestampTp0 = timestampOffsets.get(new TopicPartition(topic, 0))
-    assertEquals(0, timestampTp0.offset)
-    assertEquals(0, timestampTp0.timestamp)
-    assertEquals(Optional.of(0), timestampTp0.leaderEpoch)
-
-    val timestampTp1 = timestampOffsets.get(new TopicPartition(topic, 1))
-    assertEquals(20, timestampTp1.offset)
-    assertEquals(20, timestampTp1.timestamp)
-    assertEquals(Optional.of(0), timestampTp1.leaderEpoch)
   }
 
   def changeConsumerSubscriptionAndValidateAssignment[K, V](consumer: Consumer[K, V],
